pyproject.toml
[tool.poetry]
name = "llm-assistant"
version = "0.1.0"
description = ""
authors = ["VAIT"]
readme = "README.md"
packages = [{include = "discord_bot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
litellm = {extras = ["proxy"], version = "^1.23.7"}
openai = "^1.11.1"
python-dotenv = "^1.0.1"
discord-py-interactions = "5.11.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"
ruff = "^0.2.1"
pytest = "^8.0.0"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.5"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 127
indent-width = 2

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
show_error_codes = true
warn_return_any = true
warn_unused_ignores = true

# Prevent mypy from trigger the missing import error for our written packages
[[tool.mypy.overrides]]
module = ["discord_bot.*", "src.discord_bot.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
